Move routes.js to src/apps/mydb/
